requirements.txt
dnslib
python-dotenv
discord.py
binascii
urllib
http
socketserver